Moves necessary fixtures into core
