+++ isis/src/core/tests/CMakeLists.txt
@@ -6,6 +6,7 @@ include(../cmake/gtest.cmake)
 include(GoogleTest)
 
 set(test_source ${PROJECT_SOURCE_DIR}/tests/ConstantsTests.cpp
+                ${PROJECT_SOURCE_DIR}/tests/Fixtures.cpp
                 ${PROJECT_SOURCE_DIR}/tests/IsisCoreTestMain.cpp
                 ${PROJECT_SOURCE_DIR}/tests/PvlKeywordTests.cpp
                 ${PROJECT_SOURCE_DIR}/tests/TableTests.cpp)
